requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
requests==2.31.0
boto3
google-cloud-dialogflow
aiohttp==3.8.5
pytest
pytest-asyncio
